Adds ability to preserve keys in camelize and underscorize.
Added a new class ProtectedDict. When transforming keys, both camelize
and underscorize will preserve key case in this dictionary, but will
process all nested dictionaries.

Alternatively, camelize and underscorize checks for _protected_keys
attribute. If the attibute is set on a dictionary, keys are preserved
in the same way as with ProtectedDict.

Added helper function preserve_dict_keys(dict), which either
adds the _protected_keys attribute, or converts the dictionary
to ProtectedDict.