fix bad (i.e. too strict) type checking
This resolves an issue where `ReturnDict` objects returned from DRF
serializers would not be correctly camelized, instead passing through a
response with the original unmodified snake_case keys.

Note to future editors: `type` checking is too specific here because it does
not account for class inheritance. In this case, `ReturnDict` is a subclass
of `OrderedDict`, but `type` does not recognize that.